feat(silent_zap): rename silentlogger to silenzap, add custom logger functionality
