+++ core/silentzap/logger.go
@@ -1,4 +1,4 @@
-package silentlogger
+package silentzap
 
 import (
 	"context"
